When the socket is closed, restart the watcher
